# feat(bin): display ETA with second precision