Before:
```console
2023-06-15T16:43:08.853485Z  INFO reth::cli: Status connected_peers=2 stage=Bodies checkpoint=83.3% eta=3m 54s 910ms 884us 574ns
```

After:
```console
2023-06-15T16:43:08.853485Z  INFO reth::cli: Status connected_peers=2 stage=Bodies checkpoint=83.3% eta=3m 54s
```

---

Test is not flaky:
```console
➜  reth git:(alexey/eta-display-seconds) ✗ hyperfine -w 1 -r 100 "cargo test --release --lib --package reth node::events::tests::eta_display"
Benchmark 1: cargo test --release --lib --package reth node::events::tests::eta_display
  Time (mean ± σ):     522.7 ms ±  71.7 ms    [User: 342.2 ms, System: 157.1 ms]
  Range (min … max):   477.6 ms … 1118.9 ms    100 runs
```